Refactored /utils
