+++ src/consts.js
@@ -9,7 +9,14 @@ const SortType = {
   PRICE: `price`,
 };
 
+const MINUTES_IN_HOUR = 60;
+const MINUTES_IN_DAY = 1440;
+const HOURS_IN_DAY = 24;
+
 export {
   RenderPosition,
   SortType,
+  MINUTES_IN_HOUR,
+  MINUTES_IN_DAY,
+  HOURS_IN_DAY,
 };
